# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Diner - Restaurant Operations Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/categories:
    get:
      operationId: listCategories
      tags: [menu]
      summary: List menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [menu]
      summary: Create a menu category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /menu/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [menu]
      summary: Update a category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [menu]
      summary: Delete a category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── MENU ITEMS ─────────────────────────────────────────────────
  /menu/items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: available
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menu/items/{id}/availability:
    patch:
      operationId: toggleMenuItemAvailability
      tags: [menu]
      summary: Toggle menu item availability
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityUpdate"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

  # ─── CUSTOMERS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        sortOrder:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        sortOrder:
          type: integer
    MenuItem:
      type: object
      required: [id, categoryId, name, price, available, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        available:
          type: boolean
        imageKey:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MenuItemInput:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId:
          type: integer
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        available:
          type: boolean
        imageKey:
          type: string
    MenuItemUpdate:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        price:
          type: number
          minimum: 0
        available:
          type: boolean
        imageKey:
          type: ["string", "null"]
    AvailabilityUpdate:
      type: object
      required: [available]
      properties:
        available:
          type: boolean

    # ─── CUSTOMER SCHEMAS ────────────────────────────────────────
...